Majority working, added screens.
I want this to be a mostly polished app. That means getting all of the
trival bits around the game actually working, like a menu. The menu only
has one button, but that doesn't matter. When you open the app, you are
dropped in the menu screen, then you click the button to start a game.
when the game ends you return to the menu screen where you can start
another game. That was the key component I was missing.

I'd like to also use this as a chance to figure out icons and such.
Hopefully this will be a full fledged game that I can actually play or
have some one else play and not just a proof of concept toy.